refactor: 타입 관련 내용 폴더로 분리 #14
