+++ src/types/models/user.ts
@@ -1,0 +1,9 @@
+export interface User {
+  userId: number;
+  loginId: string;
+  name: string;
+  email: string;
+  phoneNum: string;
+  introduction: string;
+  age: number;
+}
